# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v2/troom/bulletin/IoMT_status": {
            "get": {
                "description": "IoMT_status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IoMT_status"
                ],
                "summary": "IoMT_status",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v2/troom/phone/IoMT_task": {
            "post": {
                "description": "IoMT_task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IoMT_task"
                ],
                "summary": "IoMT_task",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Notify"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/heartbeat": {
            "get": {
                "description": "Heartbeat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Heartbeat"
                ],
                "summary": "Heartbeat",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "request.Notify": {
            "type": "object",
            "properties": {
                "organization": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        }
    }
}